% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_agent.R
\name{oc_agent}
\alias{oc_agent}
\title{Query Opencast for a single capture agent}
\usage{
oc_agent(agentId = "")
}
\arguments{
\item{agentId}{The maximum number of results to return for a single request.}
}
\value{
A list containing details of a single capture agent.
}
\description{
Returns a single capture agent.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

agents <- oc_list_agents()

oc_agent(agents$content$agent_id)
}
