% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_api.R
\name{oc_api}
\alias{oc_api}
\title{Query Opencast base api}
\usage{
oc_api()
}
\value{
A list containing the default api version and server url.
}
\description{
Returns key characteristics of the API such as the server name and the default version.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_api()
}
