% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{oc_api_version}
\alias{oc_api_version}
\title{Request the available version as well as the default version}
\usage{
oc_api_version(oc_con)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}
}
\description{
Returns a list of available version as well as the default version.
}
