% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{oc_api_version_default}
\alias{oc_api_version_default}
\title{Request the default version}
\usage{
oc_api_version_default(oc_con)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}
}
\description{
Returns the default version.
}
