% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{oc_base}
\alias{oc_base}
\title{Request key characteristics}
\usage{
oc_base(oc_con)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}
}
\description{
Returns key characteristics of the API such as the server name and the default version.
}
