% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{oc_event_acl}
\alias{oc_event_acl}
\title{Request an event's access policy}
\usage{
oc_event_acl(oc_con, eventid)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{eventid}{The event id}
}
\description{
Returns an event's access policy.
}
