% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{oc_event_list}
\alias{oc_event_list}
\title{Request a list of events}
\usage{
oc_event_list(oc_con, sign = FALSE, withacl = FALSE,
  withmetadata = FALSE, withpublications = FALSE, filter = "",
  sort = "", limit = 0, offset = 0)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{sign}{Whether public distribution urls should be signed.}

\item{withacl}{Whether the acl metadata should be included in the response.}

\item{withmetadata}{Whether the metadata catalogs should be included in the response.}

\item{withpublications}{Whether the publication ids and urls should be included in the response.}

\item{filter}{A comma seperated list of filters to limit the results with. A filter is the filter's name followed by a colon ":" and then the value to filter with so it is the form :.}

\item{sort}{Sort the results based upon a list of comma seperated sorting criteria. In the comma seperated list each type of sorting is specified as a pair such as: :ASC or :DESC. Adding the suffix ASC or DESC sets the order as ascending or descending order and is mandatory.}

\item{limit}{The maximum number of results to return for a single request.}

\item{offset}{Number of results to skip based on the limit. 0 is the first set of results up to the limit, 1 is the second set of results after the first limit, 2 is third set of results after skipping the first two sets of results etc.}
}
\description{
Returns a list of events. By setting the optional sign parameter to true, the method will pre-sign distribution urls if signing is turned on in Opencast. Remember to consider the maximum validity of signed URLs when caching this response.
}
