% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_publications.R
\name{oc_event_list_publications}
\alias{oc_event_list_publications}
\title{Query Opencast for a specific events publications}
\usage{
oc_event_list_publications(eventId = "", sign = FALSE)
}
\arguments{
\item{eventId}{The event id}

\item{sign}{Whether public distribution urls should be signed.}
}
\value{
Returns an event's list of publications.
}
\description{
Returns an event's list of publications.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
# Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
# Sys.setenv(OPENCAST_USERNAME = "admin")
# Sys.setenv(OPENCAST_PASSWORD = "opencast")

# oc_event_list_publications("91df4c57-9631-4a84-b814-d8e9b36a3c32")
}
