% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{oc_event_metadata}
\alias{oc_event_metadata}
\title{Request an event's metadata}
\usage{
oc_event_metadata(oc_con, eventid, type = "dublincore/episode")
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{eventid}{The event id}

\item{type}{The type of metadata to get}
}
\description{
Returns the event's metadata of the specified type. For a metadata catalog there is the flavor such as 'dublincore/episode' and this is the unique type.
}
