% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{oc_event_publication}
\alias{oc_event_publication}
\title{Request a single publication of an event}
\usage{
oc_event_publication(oc_con, eventid, publicationid)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{eventid}{The event id}

\item{publicationid}{The publication id}
}
\description{
Returns a single publication.
}
