% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{oc_event_publications_list}
\alias{oc_event_publications_list}
\title{Request an event's list of publications.}
\usage{
oc_event_publications_list(oc_con, eventid, type = "dublincore/episode")
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{eventid}{The event id}
}
\description{
Returns an event's list of publications.
}
