% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{oc_event_single}
\alias{oc_event_single}
\title{Request a single event}
\usage{
oc_event_single(oc_con, eventid, sign = FALSE, withacl = FALSE,
  withmetadata = FALSE, withpublications = FALSE)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{eventid}{The event id}

\item{sign}{Whether public distribution urls should be signed.}

\item{withacl}{Whether the acl metadata should be included in the response.}

\item{withmetadata}{Whether the metadata catalogs should be included in the response.}

\item{withpublications}{Whether the publication ids and urls should be included in the response.}
}
\description{
Returns a single event. By setting the optional sign parameter to true, the method will pre-sign distribution urls if signing is turned on in Opencast. Remember to consider the maximum validity of signed URLs when caching this response.
}
