% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_group.R
\name{oc_group}
\alias{oc_group}
\title{Query Opencast for a specific group}
\usage{
oc_group(groupId = "")
}
\arguments{
\item{groupId}{The group id}
}
\value{
Returns a single group.
}
\description{
Returns a single group.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_group("MH_DEFAULT_ORG_EXTERNAL_APPLICATIONS")
}
