% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{oc_groups_list}
\alias{oc_groups_list}
\title{Request a list of groups.}
\usage{
oc_groups_list(oc_con, filter = "", sort = "", limit = 0,
  offset = 0)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{filter}{A comma seperated list of filters to limit the results with. A filter is the filter's name followed by a colon ":" and then the value to filter with so it is the form :.}

\item{sort}{Sort the results based upon a list of comma seperated sorting criteria. In the comma seperated list each type of sorting is specified as a pair such as: :ASC or :DESC. Adding the suffix ASC or DESC sets the order as ascending or descending order and is mandatory.}

\item{limit}{The maximum number of results to return for a single request.}

\item{offset}{Number of results to skip based on the limit. 0 is the first set of results up to the limit, 1 is the second set of results after the first limit, 2 is third set of results after skipping the first two sets of results etc.}
}
\description{
Returns a list of groups.
}
