% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{oc_groups_single}
\alias{oc_groups_single}
\title{Request a single group}
\usage{
oc_groups_single(oc_con, groupid)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{groupid}{The group id}
}
\description{
Returns a single group.
}
