% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_hostname.R
\name{oc_hostname}
\alias{oc_hostname}
\title{Opencast hostname}
\usage{
oc_hostname()
}
\value{
A character string containing the Opencast hostname.
}
\description{
This function returns the Opencast hostname that is set in the environment variable OPENCAST_HOST.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")

oc_hostname()
}
\seealso{
\code{\link{oc_username}} \code{\link{oc_password}}
}
