% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_info_me.R
\name{oc_info_me}
\alias{oc_info_me}
\title{Query Opencast current user}
\usage{
oc_info_me()
}
\value{
A list containing the information about the current user.
}
\description{
Returns information on the logged in user.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_info_me()
}
