% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_info_organization.R
\name{oc_info_organization}
\alias{oc_info_organization}
\title{Query Opencast current organization}
\usage{
oc_info_organization()
}
\value{
A list containing the settings for the anonymous role, the name, the admin role and the id.
}
\description{
Returns the current organization.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_info_organization()
}
