% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_info_organization_properties.R
\name{oc_info_organization_properties}
\alias{oc_info_organization_properties}
\title{Query Opencast current organization's properties}
\usage{
oc_info_organization_properties()
}
\value{
A list containing the current organization's properties.
}
\description{
Returns the current organization's properties.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_info_organization_properties()
}
