% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_list_agents.R
\name{oc_list_agents}
\alias{oc_list_agents}
\title{Query Opencast list of capture agent}
\usage{
oc_list_agents(limit = 0, offset = 0)
}
\arguments{
\item{limit}{The maximum number of results to return for a single request.}

\item{offset}{The index of the first result to return.}
}
\value{
A list containing the capture agents
}
\description{
Returns a list of agents.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_list_agents()
}
