% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_package_query.R
\name{oc_package_query}
\alias{oc_package_query}
\title{opencastR query function}
\usage{
oc_package_query(url = "https://legacy.opencast.org/api/",
  query = NULL)
}
\arguments{
\item{url}{The URL for the api call.}

\item{query}{A list of query parameters.}
}
\value{
The response of the api call.
}
\description{
This is an internal function of the opencastR package to query the api.
}
\examples{
oc_package_query()
}
