% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_package_useragent.R
\name{oc_package_useragent}
\alias{oc_package_useragent}
\title{opencastR user agent}
\usage{
oc_package_useragent()
}
\value{
A character string containing the user agent used in api calls generate with the opencastR package.
}
\description{
This function returns the user agent used in api calls of the opencastR package.
}
\examples{
oc_package_useragent()
}
