% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_password.R
\name{oc_password}
\alias{oc_password}
\title{Opencast password}
\usage{
oc_password()
}
\value{
A character string containing the users Opencast password.
}
\description{
This function returns the users Opencast password that is set in the environment variable OPENCAST_PASSWORD.
}
\examples{
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_password()
}
\seealso{
\code{\link{oc_hostname}} \code{\link{oc_username}}
}
