% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{oc_series_properties}
\alias{oc_series_properties}
\title{Request a single series properties}
\usage{
oc_series_properties(oc_con, seriesid)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{seriesid}{The series id}
}
\description{
Returns a series' properties
}
