% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{oc_series_single}
\alias{oc_series_single}
\title{Request a single series}
\usage{
oc_series_single(oc_con, seriesid)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}

\item{seriesid}{The series id}
}
\description{
Returns a single series.
}
