% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{oc_user}
\alias{oc_user}
\title{Request information on the logged in user}
\usage{
oc_user(oc_con)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}
}
\description{
Returns information on the logged in user.
}
