% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{oc_user_roles}
\alias{oc_user_roles}
\title{Request the roles of the logged in user}
\usage{
oc_user_roles(oc_con)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}
}
\description{
Returns current user's roles.
}
