% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_username.R
\name{oc_username}
\alias{oc_username}
\title{Opencast username}
\usage{
oc_username()
}
\value{
A character string containing the Opencast username.
}
\description{
This function returns the Opencast username that is set in the environment variable OPENCAST_USERNAME.
}
\examples{
# Set the environment variable
Sys.setenv(OPENCAST_USERNAME = "admin")

# Read out the environment variable
oc_username()
}
\seealso{
\code{\link{oc_hostname}} \code{\link{oc_password}}
}
