% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_version.R
\name{oc_version}
\alias{oc_version}
\title{Query Opencast api versions}
\usage{
oc_version()
}
\value{
A list containing the available api versions
}
\description{
Returns a list of available version as well as the default version.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_version()
}
