% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_version_default.R
\name{oc_version_default}
\alias{oc_version_default}
\title{Query Opencast default api versions}
\usage{
oc_version_default()
}
\value{
A list containing the default api version.
}
\description{
Returns a list containing the default api version.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_version_default()
}
