% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_workflow_definition.R
\name{oc_workflow_definition}
\alias{oc_workflow_definition}
\title{Query Opencast for a specific workflow definition}
\usage{
oc_workflow_definition(workflowDefinitionId = "")
}
\arguments{
\item{workflowDefinitionId}{The workflow definition id}
}
\value{
Returns a single workflow definition.
}
\description{
Returns a single workflow definition.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_workflow_definition("schedule-and-upload")
}
