% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_workflow_instance.R
\name{oc_workflow_instance}
\alias{oc_workflow_instance}
\title{Query Opencast for a specific workflow instance}
\usage{
oc_workflow_instance(workflowInstanceId = 0)
}
\arguments{
\item{workflowInstanceId}{The workflow instance id}
}
\value{
Returns a single workflow instance.
}
\description{
Returns a single workflow instance.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

workflow_instances <- oc_list_workflow_instances()

oc_workflow_instance(workflow_instances$content$identifier[1])
}
