% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_agents.R
\name{print.oc_agents}
\alias{print.oc_agents}
\title{Print result of oc_agents()}
\usage{
\method{print}{oc_agents}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_agents() function.
}
\description{
Print a structured return of the oc_agents() function.
}
\examples{
# Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
# Sys.setenv(OPENCAST_USERNAME = "admin")
# Sys.setenv(OPENCAST_PASSWORD = "opencast")

# resp <- oc_agents()

# resp
}
\seealso{
\code{\link{oc_agents}}
}
