% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_api.R
\name{print.oc_api}
\alias{print.oc_api}
\title{Print result of oc_api()}
\usage{
\method{print}{oc_api}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_api() function.
}
\description{
Print a structured return of the oc_api() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_api()

resp
}
\seealso{
\code{\link{oc_api}}
}
