% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event.R
\name{print.oc_event}
\alias{print.oc_event}
\title{Print result of oc_event()}
\usage{
\method{print}{oc_event}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_event() function.
}
\description{
Print a structured return of the oc_event() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_event("91df4c57-9631-4a84-b814-d8e9b36a3c32")

resp
}
\seealso{
\code{\link{oc_event}}
}
