% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_publication.R
\name{print.oc_event_publication}
\alias{print.oc_event_publication}
\title{Print result of oc_event_publication()}
\usage{
\method{print}{oc_event_publication}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_event_publication() function.
}
\description{
Print a structured return of the oc_event_publication() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_event_publication("91df4c57-9631-4a84-b814-d8e9b36a3c32",
                             "7964bdd6-b027-4bf9-91c1-5dec11af4bd5")

resp
}
\seealso{
\code{\link{oc_event_publication}}
}
