% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_events.R
\name{print.oc_events}
\alias{print.oc_events}
\title{Print result of oc_events()}
\usage{
\method{print}{oc_events}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_events() function.
}
\description{
Print a structured return of the oc_events() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_events()

resp
}
\seealso{
\code{\link{oc_events}}
}
