% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_info_organization.R
\name{print.oc_info_organization}
\alias{print.oc_info_organization}
\title{Print result of oc_info_organization()}
\usage{
\method{print}{oc_info_organization}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_info_organization() function.
}
\description{
Print a structured return of the oc_info_organization() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_info_organization()

resp
}
\seealso{
\code{\link{oc_info_organization}}
}
