% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_list_groups.R
\name{print.oc_list_groups}
\alias{print.oc_list_groups}
\title{Print result of oc_list_groups()}
\usage{
\method{print}{oc_list_groups}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_list_groups() function.
}
\description{
Print a structured return of the oc_list_groups() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_list_groups()

resp
}
\seealso{
\code{\link{oc_list_groups}}
}
