% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_organization_properties.R
\name{print.oc_organization_properties}
\alias{print.oc_organization_properties}
\title{Print result of oc_organization_properties()}
\usage{
\method{print}{oc_organization_properties}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_organization_properties() function.
}
\description{
Print a structured return of the oc_organization_properties() function.
}
\examples{
# Set the environment variables
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

# Get the API key characteristics
resp <- oc_organization_properties()

# Print out the key characteristics
resp
}
\seealso{
\code{\link{oc_organization_properties}}
}
