% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_series_metadata.R
\name{print.oc_series_metadata}
\alias{print.oc_series_metadata}
\title{Print result of oc_series_metadata()}
\usage{
\method{print}{oc_series_metadata}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_series_metadata() function.
}
\description{
Print a structured return of the oc_series_metadata() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_series_metadata("91df4c57-9631-4a84-b814-d8e9b36a3c32")

resp
}
\seealso{
\code{\link{oc_series_metadata}}
}
