% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_version_default.R
\name{print.oc_version_default}
\alias{print.oc_version_default}
\title{Print result of oc_version_default()}
\usage{
\method{print}{oc_version_default}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_version_default() function.
}
\description{
Print a structured return of the oc_version_default() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_version_default()

resp
}
\seealso{
\code{\link{oc_version_default}}
}
