% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_workflow_definition.R
\name{print.oc_workflow_definition}
\alias{print.oc_workflow_definition}
\title{Print result of oc_workflow_definition()}
\usage{
\method{print}{oc_workflow_definition}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_workflow_definition() function.
}
\description{
Print a structured return of the oc_workflow_definition() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_workflow_definition("schedule-and-upload")

resp
}
\seealso{
\code{\link{oc_workflow_definition}}
}
