% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_workflow_instance.R
\name{print.oc_workflow_instance}
\alias{print.oc_workflow_instance}
\title{Print result of oc_workflow_instance()}
\usage{
\method{print}{oc_workflow_instance}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_workflow_instance() function.
}
\description{
Print a structured return of the oc_workflow_instance() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

workflow_instances <- oc_list_workflow_instances()

resp <- oc_workflow_instance(workflow_instances$content$identifier[1])

resp
}
\seealso{
\code{\link{oc_workflow_instance}}
}
