% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_acl.R
\name{oc_event_acl}
\alias{oc_event_acl}
\title{Query Opencast for a specific events acl}
\usage{
oc_event_acl(eventId = "")
}
\arguments{
\item{eventId}{The event id}
}
\value{
Returns an event's access policy.
}
\description{
Returns an event's access policy.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

events <- oc_list_events()

oc_event_acl(events$content$identifier[1])
}
