% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_list_publications.R
\name{oc_event_list_publications}
\alias{oc_event_list_publications}
\title{Query Opencast for a specific events publications}
\usage{
oc_event_list_publications(eventId = "", sign = FALSE)
}
\arguments{
\item{eventId}{The event id}

\item{sign}{Whether public distribution urls should be signed.}
}
\value{
Returns an event's list of publications.
}
\description{
Returns an event's list of publications.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

events <- oc_list_events()

oc_event_list_publications(events$content$identifier[1])
}
