% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_metadata.R
\name{oc_event_metadata}
\alias{oc_event_metadata}
\title{Query Opencast for a specific events metadata}
\usage{
oc_event_metadata(eventId = "", type = "")
}
\arguments{
\item{eventId}{The event id}

\item{type}{The type of metadata to get}
}
\value{
Returns the event's metadata of the specified type.
}
\description{
Returns the event's metadata of the specified type.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
# Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
# Sys.setenv(OPENCAST_USERNAME = "admin")
# Sys.setenv(OPENCAST_PASSWORD = "opencast")

# oc_event_metadata("91df4c57-9631-4a84-b814-d8e9b36a3c32")
}
