% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_scheduling.R
\name{oc_event_scheduling}
\alias{oc_event_scheduling}
\title{Query Opencast for a specific events scheduling}
\usage{
oc_event_scheduling(eventId = "")
}
\arguments{
\item{eventId}{The event id}
}
\value{
Returns an event's scheduling information.
}
\description{
Returns an event's scheduling information.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
\donttest{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

events <- oc_list_events()

oc_event_scheduling(events$content$identifier[1])
}
}
