% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{oc_organization_properties}
\alias{oc_organization_properties}
\title{Request information about the current organization's properties}
\usage{
oc_organization_properties(oc_con)
}
\arguments{
\item{oc_con}{A vector consiting of the elements: Opencast URL, user and password.}
}
\description{
Returns the current organization's properties.
}
