% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_series.R
\name{oc_series}
\alias{oc_series}
\title{Query Opencast for a specific series}
\usage{
oc_series(seriesId = "")
}
\arguments{
\item{seriesId}{The series id}
}
\value{
Returns a single series.
}
\description{
Returns a single series.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

series <- oc_list_series()

oc_series(series$content$identifier[1])
}
