% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_series_properties.R
\name{oc_series_properties}
\alias{oc_series_properties}
\title{Query Opencast for a specific series properties}
\usage{
oc_series_properties(seriesId = "")
}
\arguments{
\item{seriesId}{The event id}
}
\value{
Returns a series' properties.
}
\description{
Returns a series' properties.
}
\details{
This function expects the hostname, username and password to be set as environment variables.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

series <- oc_list_series()

oc_series_properties(series$content$identifier[1])
}
