% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_acl.R
\name{print.oc_event_acl}
\alias{print.oc_event_acl}
\title{Print result of oc_event_acl()}
\usage{
\method{print}{oc_event_acl}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_event_acl() function.
}
\description{
Print a structured return of the oc_event_acl() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

events <- oc_list_events()

resp <- oc_event_acl(events$content$identifier[1])

resp
}
\seealso{
\code{\link{oc_event_acl}}
}
