% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_event_publication.R
\name{print.oc_event_publication}
\alias{print.oc_event_publication}
\title{Print result of oc_event_publication()}
\usage{
\method{print}{oc_event_publication}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_event_publication() function.
}
\description{
Print a structured return of the oc_event_publication() function.
}
\examples{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

events <- oc_list_events()
publications <- oc_event_list_publications(events$content$identifier[1])

resp <- oc_event_publication(events$content$identifier[1],
                             publications$content$id[1])

resp
}
\seealso{
\code{\link{oc_event_publication}}
}
