% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_series_acl.R
\name{print.oc_series_acl}
\alias{print.oc_series_acl}
\title{Print result of oc_series_acl()}
\usage{
\method{print}{oc_series_acl}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the oc_series_acl() function.
}
\description{
Print a structured return of the oc_series_acl() function.
}
\examples{
\donttest{
Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

series <- oc_list_series()

resp <- oc_series_acl(series$content$identifier[1])

resp
}
}
\seealso{
\code{\link{oc_series_acl}}
}
